completed CRUD Products

+++ Product CRUD/Program.cs
@@ -22,7 +22,6 @@ while (loop == true)
     Console.ResetColor();
 
     string choice = (Console.ReadLine());
-    
     Console.Clear();
 
     switch (choice)
@@ -34,38 +33,87 @@ while (loop == true)
             Console.BackgroundColor = ConsoleColor.DarkBlue;
             Console.WriteLine("|-----Create Product----|\n");
             Console.ResetColor();
+            createName:
             Console.Write("Enter product's Name: ");
             string name1 = Console.ReadLine();
+            if (name1 == "")
+            {
+                Console.WriteLine("Press any key to re-enter");
+                goto createName;
+            }
+            createDescription:
             Console.Write("Enter product's Description: ");
             string description1 = Console.ReadLine();
+            if (description1 == "")
+            {
+                Console.WriteLine("Press any key to re-enter");
+                goto createDescription;
+            }
             Console.WriteLine("-------------------------\n");
 
             CreateProduct(name1, description1);
             break;
         case "3":
+            
+            int id = 0;
+            string name = string.Empty;
+            string description = string.Empty;
+
             Console.BackgroundColor = ConsoleColor.DarkBlue;
             Console.WriteLine("|-----Update Product----|\n");
             Console.ResetColor();
+            updateId:
             Console.Write("Enter product's ID: ");
-            int id1 = int.Parse(Console.ReadLine());
+            string[] values = new string[] { "1", "2", "3", "4", "5", "6", "7", "8", "9,", "0" };
+            var getUpdate = Console.ReadLine();
+            if (getUpdate != "" && (values.Any(getUpdate.Contains)))
+            {
+                id = int.Parse(getUpdate);
+            }
+            else
+            {
+                Console.WriteLine("Press any key to re-enter");
+                goto updateId;
+            }
+            updateName:
             Console.Write("Enter product's Name: ");
-            string name2 = Console.ReadLine();
+            name = Console.ReadLine();
+            if (name == "")
+            {
+                Console.WriteLine("Press any key to re-enter");
+                goto updateName;
+            }
+            updateDescription:
             Console.Write("Enter product's Description: ");
-            string description2 = Console.ReadLine();
+            description = Console.ReadLine();
+            if (description == "")
+            {
+                Console.WriteLine("Press any key to re-enter");
+                goto updateDescription;
+            }
             Console.WriteLine("-------------------------\n");
 
-            UpdateProduct(id1 ,name2, description2);
+            UpdateProduct(id ,name, description);
             break;
         case "4":
             Console.BackgroundColor = ConsoleColor.DarkBlue;
             Console.WriteLine("|-----Delete Product----|\n");
             Console.ResetColor();
+            deleteId:
             Console.Write("Enter product's ID: ");
-            int id2 = int.Parse(Console.ReadLine());
+            string[] values1 = new string[] {"1","2","3","4","5","6","7","8","9,","0"};
+            var getDelete = Console.ReadLine();
+            if (getDelete != "" && (values1.Any(getDelete.Contains))) 
+            {
+                int id2 = int.Parse(getDelete);
+                DeleteProduct(id2);
+            }
+            else
+            {
+                Console.WriteLine("Press any key to re-enter");
+                goto deleteId;
+            }
             Console.WriteLine("-------------------------\n");
-
-            DeleteProduct(id2);
-
             break;
         case "5":
             Console.BackgroundColor = ConsoleColor.DarkGreen;
@@ -74,6 +122,7 @@ while (loop == true)
             loop = false;
             break;
         default:
+
             Console.BackgroundColor = ConsoleColor.DarkRed;
             Console.WriteLine("Invalid input.");
             Console.ResetColor();
@@ -91,6 +140,7 @@ void CreateProduct(string name, string description)
 
     products.Add(product);
 
+    Console.Clear();
     Console.WriteLine("----------------Created Product----------------\n");
     Console.WriteLine($"Product's ID: {product.Id}");
     Console.WriteLine($"Product's Name: {product.Name}");
@@ -101,6 +151,8 @@ void GetAllProducts()
 {
     if (products.Count==0)
     {
+        Console.Clear();
+
         Console.WriteLine($"-----------------------------GetAllProducts------------------------------\n");
         Console.WriteLine("No product's here, you should create them");
         Console.WriteLine("--------------------------------------------------------------------------\n");
@@ -108,6 +160,7 @@ void GetAllProducts()
     }
     else
     {
+        Console.Clear();
         Console.WriteLine($"--------------------------GetAllProducts---------------------------------\n");
         foreach (var product in products)
         {
@@ -127,6 +180,7 @@ void DeleteProduct(int id)
     {
         if (product.Id == id)
         {
+            Console.Clear();
             Console.WriteLine("------------------Delete Product-----------------\n");
             Console.WriteLine($"Product with ID {product.Id} removed");
             Console.WriteLine("-------------------------------------------------\n");
@@ -137,6 +191,7 @@ void DeleteProduct(int id)
     }
     if (!found)
     {
+        Console.Clear();
         Console.WriteLine("------------------Delete Product-----------------\n");
         Console.WriteLine("no product with such an ID");
         Console.WriteLine("-------------------------------------------------\n");
@@ -158,6 +213,7 @@ void UpdateProduct(int id, string name, string description)
             {
                 product.Description = description;
             }
+            Console.Clear();
             Console.WriteLine("------------------Update Product-----------------\n");
             Console.WriteLine($"Product with ID {product.Id} updated");
             Console.WriteLine("-------------------------------------------------\n");
@@ -167,6 +223,7 @@ void UpdateProduct(int id, string name, string description)
     }
     if (!found)
     {
+        Console.Clear();
         Console.WriteLine("------------------Update Product-----------------\n");
         Console.WriteLine("no product with such an ID");
         Console.WriteLine("-------------------------------------------------\n");
